# backend/users/src/main/java/com/fernando/apps/helpdeskng/users/entity/User.java
package com.fernando.apps.helpdeskng.users.entity;

import java.util.Date;
import javax.persistence.Embedded;

import javax.xml.bind.annotation.XmlAccessorType;
import javax.xml.bind.annotation.XmlRootElement;
import javax.persistence.Entity;
import javax.persistence.EnumType;
import javax.persistence.Enumerated;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Table;
import javax.validation.constraints.NotNull;
import javax.xml.bind.annotation.XmlAccessType;

/**
 * User
 */
@Entity
@Table(name = "users")
public class User {
    @Id
    @GeneratedValue(strategy=GenerationType.IDENTITY)
    private Long id;
    @NotNull
    private String name;
    @Enumerated(EnumType.STRING)
    ApplicationRoles role;
    Date createdDate;
    @Embedded
    private Credential credential;

    public Credential getCredential() {
        return credential;
    }

    public void setCredential(Credential credential) {
        this.credential = credential;
    }
    public User(){}

    public User(Long id, String name, ApplicationRoles role){
        this.id=id;
        this.name=name;
        this.role=role;
        this.createdDate=new Date();
    }

    /**
     * @return the id
     */
    public Long getId() {
        return id;
    }

    /**
     * @param id the id to set
     */
    public void setId(Long id) {
        this.id = id;
    }

    /**
     * @return the name
     */
    public String getName() {
        return name;
    }

    /**
     * @param name the name to set
     */
    public void setName(String name) {
        this.name = name;
    }

    /**
     * @return the role
     */
    public ApplicationRoles getRole() {
        return role;
    }

    /**
     * @param role the role to set
     */
    public void setRole(ApplicationRoles role) {
        this.role = role;
    }

    /**
     * @return the createdDate
     */
    public Date getCreatedDate() {
        return createdDate;
    }

    /**
     * @param createdDate the createdDate to set
     */
    public void setCreatedDate(Date createdDate) {
        this.createdDate = createdDate;
    }


}